requirements.txt
flask
flask-socketio
eventlet
google-generativeai
python-dotenv